+++ src/components/article-params-form/ArticleParamsForm.module.scss
@@ -4,6 +4,7 @@
 	width: 616px;
 	height: 100%;
 	overflow: scroll;
+	overflow-x: hidden;
 	transform: translate(-616px);
 	transition: transform 0.5s ease;
 }
@@ -16,6 +17,8 @@
 	display: flex;
 	flex-shrink: 0;
 	flex-direction: column;
+	justify-content: flex-start;
+	gap: 50px;
 	box-sizing: border-box;
 	width: 616px;
 	height: auto;
